Refactor metal blit and compute command encoders
